refactor: move files into folders auth and hotel for clearer repository/project structure
