+++ overlook_hotel/src/main/java/com/example/overlook_hotel/model/entity/Room.java
@@ -1,9 +1,12 @@
-package com.example.overlook_hotel.model;
+package com.example.overlook_hotel.model.entity;
 
 import jakarta.persistence.*;
 import lombok.*;
 
 import java.time.Instant;
+
+import com.example.overlook_hotel.model.enums.RoomStatus;
+
 import java.math.BigDecimal;
 
 @Entity
