+++ overlook_hotel/src/main/java/com/example/overlook_hotel/service/hotel/NotificationService.java
@@ -1,0 +1,5 @@
+package com.example.overlook_hotel.service.hotel;
+
+public class NotificationService {
+    
+}
